Incr 8 - Конфигурации вынес в отдельные пакеты
